refactor(inventory): 발주 카트를 PurchaseOrderCartService 로 분리 (#724 항목 5-b)
카트 CRUD 5개 + private getCartItemById 를 옮겼다. 메서드 본문 변경 0줄 — 잘라내기와
배선만. 단일 service 로 둔다(검증이 "내 것인가" 하나뿐이라 manager/reader 로 나눌
실익이 없다).

createPurchaseOrderFromCart 는 PurchaseOrderService 에 남는다 — 카트를 읽어 발주를
만드는 것을 한 트랜잭션으로 해야 해서 옮길 수 없다. 카트 테이블을 읽는 곳이 두
파일이 되는 건 의도이며 새 파일 docstring 이 그 이유를 적는다.

PurchaseOrderService: 971 → 733줄.

검증: type-check 0 · eslint 0 · jest 489 suite · nest build 성공
· 통합 8 suite/12 test(기준선과 동일)
